requirements.txt
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
scipy>=1.11
cryptography>=41.0
pandas>=2.0
pytest>=7.4
